Extract superclass for View and Eval commands
View and Eval share some behaviour, so extract a shared InspectCommand
superclass to contain this.
