+++ src/commands/eval.rb
@@ -1,11 +1,11 @@
 
 # frozen_string_literal: true
 
-require 'json'
+require 'command_helpers/inspect_command'
 
 module Metalware
   module Commands
-    class Eval < CommandHelpers::BaseCommand
+    class Eval < CommandHelpers::InspectCommand
       private
 
       attr_reader :command
@@ -16,25 +16,6 @@ module Metalware
 
       def run
         pretty_print_json(alces.instance_eval(command).to_json)
-      end
-
-      def pretty_print_json(json)
-        # Delegate pretty printing with colours to `jq`.
-        Open3.popen2(jq_command) do |stdin, stdout|
-          stdin.write(json)
-          stdin.close
-          puts stdout.read
-        end
-      end
-
-      def jq_command
-        "jq . #{colourize_output? ? '--color-output' : ''}"
-      end
-
-      def colourize_output?
-        # Should colourize the output if we have been forced to do so or we are
-        # outputting to a terminal.
-        options.color_output || STDOUT.isatty
       end
     end
   end
